+++ AdlsDotNetSDKUnitTest/SerializationTest.cs
@@ -139,5 +139,21 @@ namespace Microsoft.Azure.DataLake.Store.UnitTest
             VerifyFileStatus(entry.FileStatuses.FileStatus[2], "Test03", 0, DirectoryEntryType.FILE, DirectoryEntry.GetDateTimeFromServerTime(1531515372559), null, null, null, null, "owner3");
         }
 
+        [TestMethod]
+        public void TestListFileStatusSerialization3()
+        {
+            string liststatusOutput = "{\"FileStatuses\":{\"FileStatus\":[{\"length\":0,\"pathSuffix\":\"Test01\",\"type\":\"DIRECTORY\",\"blockSize\":0,\"accessTime\":-922337203685477,\"modificationTime\":-922337203685477,\"msExpirationTime\":1528320362391,\"replication\":0,\"permission\":\"770\",\"owner\":\"owner1\",\"group\":\"ownergroup1\",\"aclBit\":true}," +
+                "{\"length\":0,\"pathSuffix\":\"Test02\",\"type\":\"DIRECTORY\",\"blockSize\":0,\"accessTime\":1531515372559,\"modificationTime\":1531523888360,\"replication\":0,\"owner\":\"owner2\",\"group\":\"ownergroup2\",\"aclBit\":true,\"attributes\":[\"Share\",\"PartOfShare\"]}]}}";
+            var jsonSettings = new JsonSerializerSettings();
+            jsonSettings.Context = new System.Runtime.Serialization.StreamingContext(System.Runtime.Serialization.StreamingContextStates.All, "ParentDir");
+
+            var entry = JsonCustomConvert.DeserializeObject<DirectoryEntryListResult<DirectoryEntry>>(new MemoryStream(Encoding.UTF8.GetBytes(liststatusOutput)), jsonSettings);
+            Assert.IsNotNull(entry);
+            Assert.IsNotNull(entry.FileStatuses.FileStatus);
+            Assert.IsTrue(entry.FileStatuses.FileStatus.Count == 2);
+            VerifyFileStatus(entry.FileStatuses.FileStatus[0], "Test01", 0, DirectoryEntryType.DIRECTORY, null, null, DirectoryEntry.GetDateTimeFromServerTime(1528320362391), "770", "ownergroup1", "owner1");
+            VerifyFileStatus(entry.FileStatuses.FileStatus[1], "Test02", 0, DirectoryEntryType.DIRECTORY, DirectoryEntry.GetDateTimeFromServerTime(1531515372559), DirectoryEntry.GetDateTimeFromServerTime(1531523888360), null, null, "ownergroup2", "owner2");
+        }
+
     }
 }